+++ lib/models/cart.dart
@@ -2,7 +2,7 @@ class Cart {
   String imageUrl;
   String product;
   int quantity;
-  int price;
+  double price;
 
   Cart({required this.imageUrl, required this.product, required this.quantity, required this.price});
 
@@ -11,7 +11,7 @@ class Cart {
       imageUrl: json['product']['imageUrl'],
       product: json['product']['title'],
       quantity: json['quantity'],
-      price: json['product']['price']
+      price: json['product']['price'].toDouble()
     );
   }
 }
